+++ lib/2025-03-21/exercise.dart
@@ -21,39 +21,13 @@ class Book {
   String toString() {
     return 'Book{title: $title, publishDate: $publishDate, comment: $comment}';
   }
+
+  Book copyWith({String? title, String? comment, DateTime? publishDate}) {
+    return Book(
+      title: title ?? this.title,
+      comment: comment ?? this.comment,
+      publishDate: publishDate ?? this.publishDate,
+    );
+  }
 }
 
-void main() {
-  Book book1 = Book(
-    title: '한국책',
-    comment: '댓글1',
-    publishDate: DateTime(24, 01, 01),
-  );
-
-  Book book2 = Book(
-    title: '한국책',
-    comment: '댓글2',
-    publishDate: DateTime(24, 01, 01),
-  );
-  print(book1 == book2);
-
-  Book book3 = Book(
-    title: '미국책',
-    comment: '댓글3',
-    publishDate: DateTime(23, 03, 13),
-  );
-
-  Book book4 = Book(
-    title: '중국책',
-    comment: '댓글4',
-    publishDate: DateTime(21, 12, 25),
-  );
-
-  final books = [book1, book2, book3, book4];
-
-  final byPublishDate = books.toList()..sort((a,b) => a.publishDate.compareTo(b.publishDate));
-
-
-  print(book1 == book2);
-  print(byPublishDate);
-}
